Significant cleanup of Config.h/cc
- Split Config.h/cc into standalone/ConfigStandalone.h/cc
  Move as much standalone stuff as possible.

- Move TrackExtra class out of Track.h/cc into standalone/TrackExtra.h/cc

- Move Conformal-fit stuff into standalone as well.

- Makefiles -- distclean removes libs and executables.
