+++ RecoTracker/MkFitCore/src/Config.cc
@@ -7,10 +7,6 @@ namespace mkfit {
 
   namespace Config {
 
-    int nTracks = 10000;
-    int nEvents = 20;
-    int nItersCMSSW = 0;
-    bool loopOverFile = false;
 
     int nTotalLayers = -1;
 
@@ -19,65 +15,9 @@ namespace mkfit {
     // Multi threading and Clone engine configuration
     int numThreadsFinder = 1;
     int numThreadsEvents = 1;
-
-#if defined(__AVX512F__)
-    int numThreadsSimulation = 60;
-#else
-    int numThreadsSimulation = 12;
-#endif
-
-    int finderReportBestOutOfN = 1;
-
-    /*MM: moving out to IterationParams*/
-    //int   nlayers_per_seed = 3; // can be overriden from Geom plugin; a very confusing variable :)
     int numSeedsPerTask = 32;
 
-    // number of hits per task for finding seeds
-    int numHitsPerTask = 32;
-
-    // material effects
-    float RlgridME[Config::nBinsZME][Config::nBinsRME];
-    float XigridME[Config::nBinsZME][Config::nBinsRME];
-
-    //float chi2Cut = 15.;
-    //float chi2CutOverlap = 5.;
-    //float pTCutOverlap = 0.;
-
     seedOpts seedInput = simSeeds;
-    cleanOpts seedCleaning = noCleaning;
-
-    bool finding_requires_propagation_to_hit_pos;
-    PropagationFlags finding_inter_layer_pflags;
-    PropagationFlags finding_intra_layer_pflags;
-    PropagationFlags backward_fit_pflags;
-    PropagationFlags forward_fit_pflags;
-    PropagationFlags seed_fit_pflags;
-    PropagationFlags pca_prop_pflags;
-
-#ifdef CONFIG_PhiQArrays
-    bool usePhiQArrays = true;
-#endif
-
-    bool useCMSGeom = false;
-    bool readCmsswTracks = false;
-
-    bool dumpForPlots = false;
-    bool silent = false;
-
-    bool cf_seeding = false;
-    bool cf_fitting = false;
-
-    bool quality_val = false;
-    bool sim_val_for_cmssw = false;
-    bool sim_val = false;
-    bool cmssw_val = false;
-    bool fit_val = false;
-    bool readSimTrackStates = false;
-    bool inclusiveShorts = false;
-    bool keepHitInfo = false;
-    bool tryToSaveSimInfo = false;
-    matchOpts cmsswMatchingFW = hitBased;
-    matchOpts cmsswMatchingBK = trkParamBased;
 
     bool removeDuplicates = false;
     bool useHitsForDuplicates = true;
@@ -93,27 +33,28 @@ namespace mkfit {
     const float maxcth_ob = 1.99;  //eta 1.44
     const float maxcth_fw = 6.05;  //eta 2.5
 
-    bool useDeadModules = false;
+    // material effects
+    float RlgridME[Config::nBinsZME][Config::nBinsRME];
+    float XigridME[Config::nBinsZME][Config::nBinsRME];
 
-    bool mtvLikeValidation = false;
-    bool mtvRequireSeeds = false;
-    int cmsSelMinLayers = 12;
-    int nMinFoundHits = 10;
+    bool finding_requires_propagation_to_hit_pos;
+    PropagationFlags finding_inter_layer_pflags;
+    PropagationFlags finding_intra_layer_pflags;
+    PropagationFlags backward_fit_pflags;
+    PropagationFlags forward_fit_pflags;
+    PropagationFlags seed_fit_pflags;
+    PropagationFlags pca_prop_pflags;
 
-    bool kludgeCmsHitErrors = false;
-    bool backwardFit = false;
-    bool backwardSearch = true;
+#ifdef CONFIG_PhiQArrays
+    bool usePhiQArrays = true;
+#endif
+
+    bool useCMSGeom = false;
     bool includePCA = false;
 
-    bool json_dump_before = false;
-    bool json_dump_after = false;
+    bool silent = false;
     bool json_verbose = false;
-    std::vector<std::string> json_patch_filenames;
-    std::vector<std::string> json_load_filenames;
-    std::string json_save_iters_fname_fmt;
-    bool json_save_iters_include_iter_info_preamble = false;
 
-    void RecalculateDependentConstants() {}
   }  // namespace Config
 
 }  // end namespace mkfit
