+++ RecoTracker/MkFitCore/standalone/ConformalUtilsMPlex.cc
@@ -1,9 +1,88 @@
 #include "ConformalUtilsMPlex.h"
+#include "RecoTracker/MkFitCore/standalone/ConfigStandalone.h"
 #include "RecoTracker/MkFitCore/interface/Track.h"
 #include "RecoTracker/MkFitCore/interface/Hit.h"
 
 //#define DEBUG
-#include "Debug.h"
+#include "RecoTracker/MkFitCore/src/Debug.h"
+
+/* From MkFitter.h/.cc
+// ----------------
+  void ConformalFitTracks(bool fitting, int beg, int end);
+// ----------------
+  void MkFitter::ConformalFitTracks(bool fitting, int beg, int end) {
+    // bool fitting to determine to use fitting CF error widths
+    // in reality, this is depedent on hits used to make pulls
+    // could consider writing an array for widths for a given hit combo
+    // to give precise widths --> then would drop boolean
+    // also used to determine which hits to use
+
+    int front, middle, back;
+
+    // FIXME FITTING HITS --> assume one hit per layer and all layers found! BAD! Need vector of indices to do this right instead...
+    // can always assume 0,1,2 for seeding --> triplets in forward direction
+#ifdef INWARDFIT
+    front = (fitting ? Config::nLayers - 1
+                     : 0);  // i.e. would rather have true option not hardcoded... but set by ACTUAL last hit found
+    middle =
+        (fitting ? (Config::nLayers - 1) / 2 : 1);  // same with this one... would rather middle hit be in the middle!
+    back = (fitting ? 0 : 2);
+#else
+    front = (fitting ? 0 : 0);
+    middle = (fitting ? (Config::nLayers - 1) / 2 : 1);  // ditto above
+    back = (fitting ? Config::nLayers - 1 : 2);          // yup...
+#endif
+
+    // write to iC --> next step will be a propagation no matter what
+    conformalFitMPlex(fitting, Label, Err[iC], Par[iC], msPar[front], msPar[middle], msPar[back]);
+
+    // need to set most off-diagonal elements in unc. to zero, inflate all other elements;
+    if (fitting) {
+      using idx_t = Matriplex::idx_t;
+      const idx_t N = NN;
+#pragma omp simd
+      for (int n = 0; n < N; ++n) {
+        Err[iC].At(n, 0, 0) = Err[iC].ConstAt(n, 0, 0) * Config::blowupfit;
+        Err[iC].At(n, 0, 1) = Err[iC].ConstAt(n, 0, 1) * Config::blowupfit;
+        Err[iC].At(n, 1, 0) = Err[iC].ConstAt(n, 1, 0) * Config::blowupfit;
+        Err[iC].At(n, 1, 1) = Err[iC].ConstAt(n, 1, 1) * Config::blowupfit;
+        Err[iC].At(n, 2, 2) = Err[iC].ConstAt(n, 2, 2) * Config::blowupfit;
+        Err[iC].At(n, 3, 3) = Err[iC].ConstAt(n, 3, 3) * Config::blowupfit;
+        Err[iC].At(n, 4, 4) = Err[iC].ConstAt(n, 4, 4) * Config::blowupfit;
+        Err[iC].At(n, 5, 5) = Err[iC].ConstAt(n, 5, 5) * Config::blowupfit;
+
+        Err[iC].At(n, 0, 2) = 0.0f;
+        Err[iC].At(n, 0, 3) = 0.0f;
+        Err[iC].At(n, 0, 4) = 0.0f;
+        Err[iC].At(n, 0, 5) = 0.0f;
+        Err[iC].At(n, 1, 2) = 0.0f;
+        Err[iC].At(n, 1, 3) = 0.0f;
+        Err[iC].At(n, 1, 4) = 0.0f;
+        Err[iC].At(n, 1, 5) = 0.0f;
+        Err[iC].At(n, 2, 0) = 0.0f;
+        Err[iC].At(n, 2, 1) = 0.0f;
+        Err[iC].At(n, 2, 3) = 0.0f;
+        Err[iC].At(n, 2, 4) = 0.0f;
+        Err[iC].At(n, 2, 5) = 0.0f;
+        Err[iC].At(n, 3, 0) = 0.0f;
+        Err[iC].At(n, 3, 1) = 0.0f;
+        Err[iC].At(n, 3, 2) = 0.0f;
+        Err[iC].At(n, 3, 4) = 0.0f;
+        Err[iC].At(n, 3, 5) = 0.0f;
+        Err[iC].At(n, 4, 0) = 0.0f;
+        Err[iC].At(n, 4, 1) = 0.0f;
+        Err[iC].At(n, 4, 2) = 0.0f;
+        Err[iC].At(n, 4, 3) = 0.0f;
+        Err[iC].At(n, 4, 5) = 0.0f;
+        Err[iC].At(n, 5, 0) = 0.0f;
+        Err[iC].At(n, 5, 1) = 0.0f;
+        Err[iC].At(n, 5, 2) = 0.0f;
+        Err[iC].At(n, 5, 3) = 0.0f;
+        Err[iC].At(n, 5, 4) = 0.0f;
+      }
+    }
+  }
+*/
 
 namespace mkfit {
 
@@ -23,7 +102,7 @@ namespace mkfit {
     Tv* c = C.fArray;
     ASSUME_ALIGNED(c, 64);
 
-#include "CFMatrix33Vector3.ah"
+#include "RecoTracker/MkFitCore/standalone/CFMatrix33Vector3.ah"
   }
 
   //M. Hansroul, H. Jeremie and D. Savard, NIM A 270 (1988) 498
